changes:
added fast-xml-parser. code now fetches and parses xml data

added xmlfetcher.service.ts that contains a url for the xml data to be parsed.

on this file, zone_color and price are pulled from the zones section, then this is compared to the customfill color for each seat to assign a price.

This all has a cron job to fetch this data and then override the data in the event table in the database.